Moved campaign-page specific components into campaign app folder
